附件上传控件优化

+++ gorm-general-domain/src/main/groovy/neo/script/gorm/general/service/AttachmentService.groovy
@@ -46,9 +46,10 @@ class AttachmentService extends AbstractService<AttachmentInfo> {
 
     void deleteInfoByOwners(List ownerList) {
         log.info "deleteInfoByOwners $ownerList"
-        list(['in': [['ownerId', ownerList]]]).each { info ->
-            deleteInfo(info)
-        }
+        if (ownerList)
+            list(['in': [['ownerId', ownerList]]]).each { info ->
+                deleteInfo(info)
+            }
     }
 
     void deleteInfoByOwnerAndFileId(String ownerId, String fileId) {
